requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
cloudscraper>=1.2.71
selenium>=4.15.0

# Platform APIs
tweepy>=4.14.0
praw>=7.7.0

# NLP & Sentiment Analysis
transformers>=4.35.0
torch>=2.1.0
textblob>=0.17.1
vaderSentiment>=3.3.2
spacy>=3.7.0

# Data Processing
pandas>=2.1.0

# Configuration
pyyaml>=6.0.1

# Async & Parallelism
aiohttp>=3.9.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0

# Report Generation
markdown>=3.5.0
